% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_inverted_commas}
\alias{add_inverted_commas}
\title{Add inverted commas ' around any sheet name in Excel workbook for consistency}
\usage{
add_inverted_commas(all_sheets)
}
\arguments{
\item{all_sheets}{list of data.frames exported by tidyxl::xlsx_cells}
}
\value{
list of data.frames in format exported by tidyxl::xlsx_cells
}
\description{
Add inverted commas ' around any sheet name in Excel workbook for consistency
}
\examples{
temp_output_list <- add_inverted_commas(temp_input_list)
}
