% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_sheet_name_to_formulae}
\alias{add_sheet_name_to_formulae}
\title{Add !prefix to any formulae without a sheet !prefix}
\usage{
add_sheet_name_to_formulae(input_sheet)
}
\arguments{
\item{input_sheet}{a data.frame exported by tidyxl::xlsx_cells}

\item{workbook_name}{input_Sheet}
}
\value{
a data.frame in form exported by tidyxl::xlsx_cells
}
\description{
Add !prefix to any formulae without a sheet !prefix
}
\examples{
temp_input <- xlsx_cells("test_workbook.xlsx", sheets = sheet_name)
temp_output <- add_sheet_name_to_formulae(temp_input)
}
