% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{array_construct}
\alias{array_construct}
\title{array_construct function}
\usage{
array_construct(...)
}
\arguments{
\item{...}{vars to form the array}

\item{dims}{dimensions of the array we are constructing}
}
\value{
the constructed array
}
\description{
Correctly interprets and constructs arrays
}
\examples{
array_construct(array(c(40,40,40,40,20,20),c(2,3)),array(c(20,20,20,20,20,20),c(2,3)))
}
