% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea-interface-layer.R
\name{BCEA_all_output_loop}
\alias{BCEA_all_output_loop}
\title{Call BCEA}
\usage{
BCEA_all_output_loop(costs, effectiveness, treatments, Kmax)
}
\arguments{
\item{...}{the N inputs}
}
\value{
dataframe containing all inputs
}
\description{
Take N inputs and dumps them to screen
}
\examples{
BCEA_all_output_loop(costs,effectiveness,treatments,Kmax)

}
