% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea-interface-layer.R
\name{BCEA_dataframe}
\alias{BCEA_dataframe}
\title{Form BCEA Dataframe}
\usage{
BCEA_dataframe(...)
}
\arguments{
\item{...}{the N inputs}
}
\value{
dataframe containing all inputs
}
\description{
Take N inputs and dumps them to screen
}
\examples{
dataframe <- BCEA_dataframe(a=c(1,2,3),b=c(4,5,6))

}
