% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea-interface-layer.R
\name{BCEA_matrix}
\alias{BCEA_matrix}
\title{Form BCEA matrix}
\usage{
BCEA_matrix(...)
}
\arguments{
\item{...}{the N inputs}
}
\value{
dataframe containing all inputs
}
\description{
Take N inputs and dumps them to screen
}
\examples{
costs <- BCEA_matrix(a=c(1,2,3),b=c(4,5,6))
effectiveness <- BCEA_matrix(a=c(1,2,3),b=c(4,5,6))

}
