% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-betadist.R
\name{excel__xlfn.beta.dist}
\alias{excel__xlfn.beta.dist}
\title{Excel BETA.DIST function}
\usage{
excel__xlfn.beta.dist(x, alpha, beta, cumulative, A = 0, B = 1)
}
\arguments{
\item{x}{value to evaluate}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}

\item{A}{Optional parameter - lower bound}

\item{B}{Optional parameter - upper bound}
}
\value{
qbinom or dbinom
}
\description{
return correct norm distribution
}
\examples{
beta.dist <- excel__xlfn.beta.dist(0.2,100,6,True)

}
