% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-binom.R
\name{excel__xlfn.binom.dist}
\alias{excel__xlfn.binom.dist}
\title{Excel BINOM.DIST function}
\usage{
excel__xlfn.binom.dist(x, N, prob, cumulative)
}
\arguments{
\item{x}{Number of successful trials}

\item{N}{TOtal number of trials}

\item{prob}{Probability}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pbinom or dbinom
}
\description{
return correct norm distribution
}
\examples{
binom.dist <- excel_xlfn.binom.dist(0.2,100,6,True)

}
