% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-binom.R
\name{excel__xlfn.binom.inv}
\alias{excel__xlfn.binom.inv}
\title{Excel BINOM.INV function}
\usage{
excel__xlfn.binom.inv(N, prob, alpha)
}
\arguments{
\item{N}{Total number of trials}

\item{prob}{Probability of success}

\item{alpha}{citerion value}
}
\value{
qbinom or dbinom
}
\description{
return correct norm distribution
}
\examples{
binomdist <- excel__xlfn.binom.inv(100,0.2,6)

}
