% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel__xlfn.chisq.dist}
\alias{excel__xlfn.chisq.dist}
\title{Excel CHISQ.DIST function}
\usage{
excel__xlfn.chisq.dist(x, df, cumulative)
}
\arguments{
\item{x}{values}

\item{df}{degrees of freedom}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pchisq or dchisq
}
\description{
return correct norm distribution
}
\examples{
chisq.dist <- excel__xlfn.chisq.dist(2,100,True)

}
