% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel__xlfn.chisq.dist.rt}
\alias{excel__xlfn.chisq.dist.rt}
\title{Excel CHISQ.DIST.RT function}
\usage{
excel__xlfn.chisq.dist.rt(x, df)
}
\arguments{
\item{x}{values}

\item{df}{degrees of freedom}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
dchisq
}
\description{
return correct norm distribution
}
\examples{
chisq.dist.rt <- excel__xlfn.chisq.dist.rt(2,100)

}
