% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel__xlfn.chisq.inv.rt}
\alias{excel__xlfn.chisq.inv.rt}
\title{Excel CHISQ.INV.RT function}
\usage{
excel__xlfn.chisq.inv.rt(x, df)
}
\arguments{
\item{x}{val}

\item{df}{degrees of freedom}
}
\value{
qchisq
}
\description{
return correct norm distribution
}
\examples{
chisq.inv.rt <- excel__xlfn.chisq.inv.rt(2,100)

}
