% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel__xlfn.chisq.test}
\alias{excel__xlfn.chisq.test}
\title{Excel CHISQ.TEST function}
\usage{
excel__xlfn.chisq.test(x, y)
}
\arguments{
\item{x}{actual range}

\item{y}{expected range}
}
\value{
chisq.test
}
\description{
return correct norm distribution
}
\examples{
chisq.test<- excel__xlfn.chisq.test(2,100)

}
