% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-string.R
\name{excel__xlfn.concat}
\alias{excel__xlfn.concat}
\title{Convert CONCAT}
\usage{
excel__xlfn.concat(...)
}
\arguments{
\item{...}{up to 256 indiviual cells to be concatenated}
}
\description{
Convert CONCAT
}
