% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-expdist.R
\name{excel__xlfn.expon.dist}
\alias{excel__xlfn.expon.dist}
\title{Excel EXPON.DIST function}
\usage{
excel__xlfn.expon.dist(x, lambda, cumulative)
}
\arguments{
\item{x}{value}

\item{lambda}{scaling}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pexp or dexp
}
\description{
return correct norm distribution
}
\examples{
expon.dist <- excel__xlfn.expon.dist(0.2,100,True)

}
