% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-fdist.R
\name{excel__xlfn.f.dist}
\alias{excel__xlfn.f.dist}
\title{Excel F.DIST function}
\usage{
excel__xlfn.f.dist(x, df1, df2, cumulative)
}
\arguments{
\item{x}{value}

\item{df1}{degree of freedom 1}

\item{df2}{degree of freedom 2}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pf or df
}
\description{
return correct norm distribution
}
\examples{
f.dist <- excel__xlfn.f.dist(0.2,10,10,True)

}
