% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-fdist.R
\name{excel__xlfn.f.dist.rt}
\alias{excel__xlfn.f.dist.rt}
\title{Excel F.DIST.RT function}
\usage{
excel__xlfn.f.dist.rt(x, df1, df2)
}
\arguments{
\item{x}{value}

\item{df1}{degree of freedom 1}

\item{df2}{degree of freedom 2}
}
\value{
pf
}
\description{
return correct norm distribution
}
\examples{
f.dist.rt <- excel__xlfn.f.dist.rt(0.2,10,10)

}
