% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-fdist.R
\name{excel__xlfn.f.inv}
\alias{excel__xlfn.f.inv}
\title{Excel F.INV function}
\usage{
excel__xlfn.f.inv(x, df1, df2)
}
\arguments{
\item{x}{probability}

\item{df1}{degree of freedom 1}

\item{df2}{degree of freedom 2}
}
\value{
qf
}
\description{
return correct norm distribution
}
\examples{
f.inv <- excel__xlfn.f.inv(0.2,10,10)

}
