% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-fdist.R
\name{excel__xlfn.f.inv.rt}
\alias{excel__xlfn.f.inv.rt}
\title{Excel F.INV.RT function}
\usage{
excel__xlfn.f.inv.rt(x, df1, df2)
}
\arguments{
\item{x}{probability}

\item{df1}{degree of freedom 1}

\item{df2}{degree of freedom 2}
}
\value{
qf
}
\description{
return correct norm distribution
}
\examples{
f.inv.rt <- excel__xlfn.f.inv.rt(0.2,10,10)

}
