% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-gamma.R
\name{excel__xlfn.gamma.dist}
\alias{excel__xlfn.gamma.dist}
\title{Excel GAMMA.DIST function}
\usage{
excel__xlfn.gamma.dist(x, alpha, beta, cumulative)
}
\arguments{
\item{x}{list of random numbers}

\item{alpha}{shape in R}

\item{beta}{scale in R}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qgamma or dgamma
}
\description{
Excel GAMMA.DIST function
}
\examples{
gamma.dist <- excel__xlfn.gamma.dist(0.2,100,6,True)

}
