% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-gamma.R
\name{excel__xlfn.gamma.inv}
\alias{excel__xlfn.gamma.inv}
\title{Excel GAMMA.INV function}
\usage{
excel__xlfn.gamma.inv(x, alpha, beta)
}
\arguments{
\item{x}{list of random numbers}

\item{alpha}{shape in R}

\item{beta}{scale in R}
}
\value{
qgamma
}
\description{
Excel GAMMA.INV function
}
\examples{
gamma.inv <- excel__xlfn.gamma.inv(0.2,100,6)

}
