% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-lognormdist.R
\name{excel__xlfn.lognorm.dist}
\alias{excel__xlfn.lognorm.dist}
\title{Excel LOGNORM.DIST function}
\usage{
excel__xlfn.lognorm.dist(x, mean, sd, cumulative)
}
\arguments{
\item{x}{list of random numbers}

\item{mean}{mean of distribution}

\item{sd}{standard deviation of distribution}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qlnorm or dlnorm
}
\description{
return correct norm distribution
}
\examples{
lnorm.dist <- excel__xlfn.lnorm.dist(0.2,100,6,True)

}
