% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-binom.R
\name{excel__xlfn.negbinom.dist}
\alias{excel__xlfn.negbinom.dist}
\title{Excel NEGBINOM.DIST function}
\usage{
excel__xlfn.negbinom.dist(num_f, num_s, prob, cumulative)
}
\arguments{
\item{num_f}{Number of failed trials}

\item{num_s}{Number of successful trials}

\item{prob}{Probability}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pnbinom or dnbinom
}
\description{
return correct norm distribution
}
\examples{
negbinom.dist <- excel__xlfn.negbinom.dist(3,4,0.25,True)

}
