% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-normdist.R
\name{excel__xlfn.norm.dist}
\alias{excel__xlfn.norm.dist}
\title{Excel NORM.DIST function}
\usage{
excel__xlfn.norm.dist(x, mean, sd, cumulative)
}
\arguments{
\item{x}{list of random numbers}

\item{mean}{mean of distribution}

\item{sd}{standard deviation of distribution}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qnorm or dnorm
}
\description{
return correct norm distribution
}
\examples{
norm.dist <- excel__xlfn.norm.dist(0.2,100,6,True)

}
