% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-normdist.R
\name{excel__xlfn.norm.s.dist}
\alias{excel__xlfn.norm.s.dist}
\title{Excel NORM.S.DIST function}
\usage{
excel__xlfn.norm.s.dist(x, cumulative)
}
\arguments{
\item{x}{distribution value}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qnorm or dnorm
}
\description{
Excel NORM.S.DIST function
}
\examples{
norm.s.dist <- excel__xlfn.norm.s.dist(0.2,100,6,True)

}
