% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-poissondist.R
\name{excel__xlfn.poisson.dist}
\alias{excel__xlfn.poisson.dist}
\title{Excel POISSON.DIST function}
\usage{
excel__xlfn.poisson.dist(x, mean, cumulative)
}
\arguments{
\item{x}{Number of events}

\item{mean}{expectation value}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
ppois or dpois
}
\description{
return correct norm distribution
}
\examples{
poisson.dist <- excel__xlfn.poisson.dist(10,10,True)

}
