% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel__xlfn.quartile.inc}
\alias{excel__xlfn.quartile.inc}
\title{Convert QUARTILE.INC}
\usage{
excel__xlfn.quartile.inc(range, quart)
}
\arguments{
\item{range}{list of values}

\item{quart}{list of values}
}
\value{
returns the quartile determined by quart, with [0,1]
}
\description{
Convert QUARTILE.INC
}
\examples{
quartile.incListVars <- excel__xlfn.quartile.inc(list(2, 3, 5, 2.5, -10),1)
quartile.incMultiListVars <- excel__xlfn.quartile.inc(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)),1)

}
