% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel__xlfn.stdev.p}
\alias{excel__xlfn.stdev.p}
\title{Convert STDEV.P(...)}
\usage{
excel__xlfn.stdev.p(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
populations tandard deviation
}
\description{
Convert STDEV.P(...)
}
\examples{
stdev.pSingleVars <- excel__xlfn.stdev.p(0, 3, 5, 2.5, -10)
stdev.pListVars <- excel__xlfn.stdev.p(list(2, 3, 5, 2.5, -10))
stdev.pMultiListVars <- excel__xlfn.stdev.p(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
