% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel__xlfn.stdev.s}
\alias{excel__xlfn.stdev.s}
\title{Convert STDEV.S(...)}
\usage{
excel__xlfn.stdev.s(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
sample standard deviation
}
\description{
Convert STDEV.S(...)
}
\examples{
stdev.sSingleVars <- excel__xlfn.stdev.s(0, 3, 5, 2.5, -10)
stdev.sListVars <- excel__xlfn.stdev.s(list(2, 3, 5, 2.5, -10))
stdev.sMultiListVars <- excel__xlfn.stdev.s(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
