% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-tdist.R
\name{excel__xlfn.t.dist.2t}
\alias{excel__xlfn.t.dist.2t}
\title{Excel T.DIST.2T function}
\usage{
excel__xlfn.t.dist.2t(x, df)
}
\arguments{
\item{x}{value}

\item{df}{degrees of freedom}
}
\value{
pt
}
\description{
return correct norm distribution
}
\examples{
t.dist.2t <- excel__xlfn.t.dist.2t(0.2,10)

}
