% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-tdist.R
\name{excel__xlfn.t.dist}
\alias{excel__xlfn.t.dist}
\title{Excel T.DIST function}
\usage{
excel__xlfn.t.dist(x, df, cumulative)
}
\arguments{
\item{x}{value}

\item{df}{degrees of freedom}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pt or dt
}
\description{
return correct norm distribution
}
\examples{
t.dist <- excel__xlfn.t.dist(0.2,10,True)

}
