% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-tdist.R
\name{excel__xlfn.t.dist.rt}
\alias{excel__xlfn.t.dist.rt}
\title{Excel T.DIST.RT function}
\usage{
excel__xlfn.t.dist.rt(x, df)
}
\arguments{
\item{x}{value}

\item{df}{degrees of freedom}
}
\value{
pt
}
\description{
return correct norm distribution
}
\examples{
t.dist.rt <- excel__xlfn.t.dist.rt(0.2,10)

}
