% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-tdist.R
\name{excel__xlfn.t.inv}
\alias{excel__xlfn.t.inv}
\title{Excel T.INV function}
\usage{
excel__xlfn.t.inv(x, df)
}
\arguments{
\item{x}{value}

\item{df}{degrees of freedom}
}
\value{
qt
}
\description{
return correct norm distribution
}
\examples{
t.inv <- excel__xlfn.t.inv(0.2,10)

}
