% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel__xlfn.var.p}
\alias{excel__xlfn.var.p}
\title{Convert VAR.P(...)}
\usage{
excel__xlfn.var.p(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
population variance of list
}
\description{
Convert VAR.P(...)
}
\examples{
var.pSingleVars <- excel__xlfn.var.p(0, 3, 5, 2.5, -10)
var.pListVars <- excel__xlfn.var.p(list(2, 3, 5, 2.5, -10))
var.pMultiListVars <- excel__xlfn.var.p(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
