% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel__xlfn.var.s}
\alias{excel__xlfn.var.s}
\title{Convert VAR.S(...)}
\usage{
excel__xlfn.var.s(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
sample variance of list
}
\description{
Convert VAR.S(...)
}
\examples{
var.sSingleVars <- excel__xlfn.var.s(0, 3, 5, 2.5, -10)
var.sListVars <- excel__xlfn.var.s(list(2, 3, 5, 2.5, -10))
var.sMultiListVars <- excel__xlfn.var.s(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
