% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-weibull.R
\name{excel__xlfn.weibull.dist}
\alias{excel__xlfn.weibull.dist}
\title{Excel WEIBULL.DIST function}
\usage{
excel__xlfn.weibull.dist(x, alpha, beta, cumulative)
}
\arguments{
\item{x}{value to evaluate}

\item{alpha}{shape}

\item{beta}{scale}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qbinom or dbinom
}
\description{
Excel WEIBULL.DIST function
}
\examples{
weibull.dist <- excel__xlfn.weibull.dist(0.2,100,6,True)

}
