% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-trig.R
\name{excel_acot}
\alias{excel_acot}
\title{Convert ACOT(a,b)}
\usage{
excel_acot(a, b)
}
\arguments{
\item{a}{value}

\item{b}{exponent}
}
\value{
a^b
}
\description{
Convert ACOT(a,b)
}
\examples{
powerSingleVars <- excel_power(2, 3)
}
