% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-trig.R
\name{excel_acsc}
\alias{excel_acsc}
\title{Convert ACSC(a,b)}
\usage{
excel_acsc(a, b)
}
\arguments{
\item{a}{value}

\item{b}{exponent}
}
\value{
a^b
}
\description{
Convert ACSC(a,b)
}
\examples{
powerSingleVars <- excel_power(2, 3)
}
