% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-replacement.R
\name{excel_array}
\alias{excel_array}
\title{Construct the correct form of array}
\usage{
excel_array(dim, ...)
}
\arguments{
\item{...}{input vars}
}
\value{
array
}
\description{
Construct the correct form of array
}
