% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-trig.R
\name{excel_atan2}
\alias{excel_atan2}
\title{Convert ATAN2(a,b)}
\usage{
excel_atan2(a, b)
}
\arguments{
\item{a}{value}

\item{b}{exponent}
}
\value{
a^b
}
\description{
Convert ATAN2(a,b)
}
\examples{
powerSingleVars <- excel_power(2, 3)
}
