% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_average}
\alias{excel_average}
\title{Convert AVERAGE(...)}
\usage{
excel_average(...)
}
\arguments{
\item{...}{list of values}
}
\value{
mean of values
}
\description{
Convert AVERAGE(...)
}
\examples{
meanSingleVars <- excel_average(0, 3, 5, 2.5, -10)
meanListVars <- excel_average(list(2, 3, 5, 2.5, -10))
meanMultiListVars <- excel_average(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))
}
