% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-betadist.R
\name{excel_betadist}
\alias{excel_betadist}
\title{Excel BETADIST function}
\usage{
excel_betadist(x, alpha, beta, A = 0, B = 1)
}
\arguments{
\item{x}{value to evaluate}

\item{A}{Optional parameter - lower bound}

\item{B}{Optional parameter - upper bound}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qbinom or dbinom
}
\description{
return correct norm distribution
}
\examples{
betadist <- excel_betadist(0.2,100,6,True)

}
