% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_ceiling}
\alias{excel_ceiling}
\title{Excel CEILING Function}
\usage{
excel_ceiling(number, significance)
}
\arguments{
\item{number}{The numeric value to round up}

\item{significance}{The multiple to which you want to round}
}
\value{
The rounded number
}
\description{
Rounds a number up to the nearest multiple of significance.
This function replicates the behavior of Excel's CEILING function.
}
\examples{
excel_ceiling(2.5, 1)      # Returns 3
excel_ceiling(26.75, 0.1)  # Returns 26.8
excel_ceiling(-2.5, -2)    # Returns -4
excel_ceiling(1.2, 1)      # Returns 2

}
