% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel_chidist}
\alias{excel_chidist}
\title{Excel CHIDIST function}
\usage{
excel_chidist(x, df)
}
\arguments{
\item{x}{values}

\item{df}{degrees of freedom}
}
\value{
pchisq
}
\description{
return correct norm distribution
}
\examples{
chidist <- excel_chidist(2,100)

}
