% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel_chiinv}
\alias{excel_chiinv}
\title{Excel CHIINV function}
\usage{
excel_chiinv(x, df)
}
\arguments{
\item{x}{probability}

\item{df}{degrees of freedom}
}
\description{
return correct norm distribution
}
\examples{
chidist <- excel_chidist(2,100)

}
