% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_chisq.R
\name{excel_chitest}
\alias{excel_chitest}
\title{Excel CHITEST function}
\usage{
excel_chitest(x, y)
}
\arguments{
\item{x}{actual range}

\item{y}{expected range}
}
\value{
chisq.test
}
\description{
return correct norm distribution
}
\examples{
chitest<- excel__xlfn.chisq.inv.rt(2,100)

}
