% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-programflow.R
\name{excel_choose}
\alias{excel_choose}
\title{Excel CHOOSE function}
\usage{
excel_choose(n, ...)
}
\arguments{
\item{n}{the nth input in the list of N inputs}

\item{...}{the N inputs. N<n is forbidden}
}
\value{
The nth input
}
\description{
Take N inputs and return the nth result.
}
\examples{
sixth <- excel_choose(6, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
tenth <- excel_choose(10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)

}
