% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-string.R
\name{excel_concatenate}
\alias{excel_concatenate}
\title{Convert CONCATENATE}
\usage{
excel_concatenate(...)
}
\arguments{
\item{...}{up to 256 indiviual cells to be concatenated}
}
\description{
Convert CONCATENATE
}
