% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_count}
\alias{excel_count}
\title{Convert COUNT}
\usage{
excel_count(...)
}
\arguments{
\item{...}{list of inputs}
}
\value{
number of cells that are numeric
}
\description{
Count the number of numerical values in  the selected cells
}
\examples{
countSingleVars <- excel_count(0, 3, 5, 'A', -10)
countListVars <- excel_count(list(2, 'V', 5, 2.5, -10))
}
