% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_counta}
\alias{excel_counta}
\title{Convert COUNTA}
\usage{
excel_counta(...)
}
\arguments{
\item{...}{list of inputs}
}
\value{
number of cells that are not NA
}
\description{
Count the number of numerical values in  the selected cells
}
\examples{
countaSingleVars <- excel_counta(0, 3, 5, 'A', -10)
countaListVars <- excel_counta(list(2, 'V', 5, 2.5, -10))
}
