% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_countif}
\alias{excel_countif}
\title{Convert COUNTIF}
\usage{
excel_countif(range, criteria)
}
\arguments{
\item{range}{range of which we will compare}
}
\value{
number of cells that satisfy the criteria
}
\description{
Count the number of numerical values in  the selected cells that satisfy the criterion
}
\examples{
excel_countif(list(2, 'V', 5, 2.5, -10), "V")  # Returns: 1
excel_countif(list(2, 'V', 5, 2.5, -10), 2)    # Returns: 1
excel_countif(list(2, 'V', 5, 2.5, -10), ">2") # Returns: 2
excel_countif(list(2, 'V', 5, 2.5, -10), "<=2.5") # Returns: 3
excel_countif(list(2, 3, 5, "Test", "val", NA), "<>3") # Returns: 5
excel_countif(list(2, 3, 5, "Test", "val", NA), "=Test") # Returns: 1
}
