% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-trig.R
\name{excel_degrees}
\alias{excel_degrees}
\title{Convert DEGREES(a,b)}
\usage{
excel_degrees(rad)
}
\arguments{
\item{rad}{angle in radians}
}
\value{
angle in degrees
}
\description{
Convert DEGREES(a,b)
}
\examples{
degrees <- excel_degrees(pi)
}
