% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-expdist.R
\name{excel_expondist}
\alias{excel_expondist}
\title{Excel EXPONDIST function}
\usage{
excel_expondist(x, lambda, cumulative)
}
\arguments{
\item{x}{value}

\item{lambda}{scaling}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pexp or dexp
}
\description{
return correct norm distribution
}
\examples{
expondist <- excel_expondist(0.2,100,True)

}
