% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_floor}
\alias{excel_floor}
\title{Excel FLOOR Function}
\usage{
excel_floor(number, significance)
}
\arguments{
\item{number}{The numeric value to round down}

\item{significance}{The multiple to which you want to round}
}
\value{
The rounded number
}
\description{
Rounds a number down to the nearest multiple of significance.
This function replicates the behavior of Excel's FLOOR function.
}
\examples{
excel_floor(3.7, 2)      # Returns 2
excel_floor(26.75, 0.1)  # Returns 26.7
excel_floor(-2.5, -2)    # Returns -2
excel_floor(1.5, 1)      # Returns 1

}
