% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-gamma.R
\name{excel_gammadist}
\alias{excel_gammadist}
\title{Excel GAMMADIST function}
\usage{
excel_gammadist(x, alpha, beta, cumulative)
}
\arguments{
\item{x}{list of random numbers}

\item{alpha}{shape in R}

\item{beta}{scale in R}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qgamma or dgamma
}
\description{
Excel GAMMADIST function
}
\examples{
gammadist <- excel_gammadist(0.2,100,6,True)

}
