% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-gamma.R
\name{excel_gammainv}
\alias{excel_gammainv}
\title{Excel GAMMAINV function}
\usage{
excel_gammainv(x, alpha, beta)
}
\arguments{
\item{x}{list of random numbers}

\item{alpha}{shape in R}

\item{beta}{scale in R}
}
\value{
qgamma
}
\description{
Excel GAMMAINV function
}
\examples{
gammainv <- excel_gammainv(0.2,100,6)

}
