% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-replacement.R
\name{excel_iferror}
\alias{excel_iferror}
\title{Excel IFERROR function}
\usage{
excel_iferror(value, value_if_error)
}
\arguments{
\item{value}{expected non-blank value}

\item{value_if_error}{value to return if blank}
}
\value{
return non-blank value
}
\description{
Check if value is blank and return non-blank value.
}
\examples{
num <- excel_iferror(6)
blank <- excel_iferror(NA)

}
