% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-replacement.R
\name{excel_ignore}
\alias{excel_ignore}
\title{Ignored an Excel function}
\usage{
excel_ignore(...)
}
\arguments{
\item{...}{the N inputs.}
}
\value{
NA
}
\description{
Take N inputs and return NA.
Inform the user that the expected function is not
implimented.
}
\examples{
ignored_function <- excel_ignore(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)

}
