% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-programflow.R
\name{excel_index}
\alias{excel_index}
\title{Excel INDEX function}
\usage{
excel_index(inputarray, row, column)
}
\arguments{
\item{inputarray}{array to be accessed}

\item{row}{row of data}

\item{column}{column of data}
}
\value{
data located at (row,column)
}
\description{
Take three inputs and return the value at location
}
\examples{
test <- excel_index(array(c(1,2,3,4),dim=c(2,2)),1,1)

}
