% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-string.R
\name{excel_left}
\alias{excel_left}
\title{Convert LEFT(str, n)}
\usage{
excel_left(str, n)
}
\arguments{
\item{str}{text string}

\item{n}{number of characters to be captured}
}
\description{
Convert LEFT(str, n)
}
\examples{
 leftstringOne<- excel_left("string",1)
 leftstringTwo<- excel_left("string",5)
 leftstringThree<- excel_left("string",3)

}
