% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_log}
\alias{excel_log}
\title{Convert LOG(x, base)}
\usage{
excel_log(x, base = 10)
}
\arguments{
\item{x}{value}

\item{base}{base of logarithm}
}
\value{
log(x,b)
}
\description{
Convert LOG(x, base)
}
\examples{
log1 <- excel_log(10)
log2 <- excel_log(16,2)

}
