% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-lognormdist.R
\name{excel_lognormdist}
\alias{excel_lognormdist}
\title{Excel LOGNORMDIST function}
\usage{
excel_lognormdist(x, mean, sd)
}
\arguments{
\item{x}{list of random numbers}

\item{mean}{mean of distribution}

\item{sd}{standard deviation of distribution}
}
\value{
qlnorm or dlnorm
}
\description{
return correct norm distribution
}
\examples{
lnormdist <- excel_lnormdist(0.2,100,6)

}
