% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_max}
\alias{excel_max}
\title{Convert MAX}
\usage{
excel_max(...)
}
\arguments{
\item{...}{list of values}
}
\value{
returns max value
}
\description{
Convert MAX
}
\examples{
maxSingleVars <- excel_max(0, 3, 5, 2.5, -10)
maxListVars <- excel_max(list(2, 3, 5, 2.5, -10))
maxMultiListVars <- excel_max(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
