% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_median}
\alias{excel_median}
\title{Convert MEDIAN(...)}
\usage{
excel_median(...)
}
\arguments{
\item{...}{list}
}
\value{
median of values
}
\description{
Convert MEDIAN(...)
}
\examples{
medianSingleVars <- excel_median(0, 3, 5, 2.5, -10)
medianListVars <- excel_median(list(0, 3, 5, 2.5, -10))
medianMultiListVars <- excel_median(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))
}
