% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-string.R
\name{excel_mid}
\alias{excel_mid}
\title{Convert MID(str, n)}
\usage{
excel_mid(str, x, n)
}
\arguments{
\item{str}{text string}

\item{x}{starting location of first char}

\item{n}{number of characters to be captured}
}
\description{
Convert MID(str, n)
}
\examples{
 midstringOne<- excel_mid("string",1,1)
 midstringTwo<- excel_mid("string",1,5)
 midstringThree<- excel_mid("string",1,6)
 midstringFour<- excel_mid("string",3,1)

}
