% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_min}
\alias{excel_min}
\title{Convert MIN}
\usage{
excel_min(...)
}
\arguments{
\item{...}{list of values}
}
\value{
returns min value
}
\description{
Convert MIN
}
\examples{
minSingleVars <- excel_min(0, 3, 5, 2.5, -10)
minListVars <- excel_min(list(2, 3, 5, 2.5, -10))
minMultiListVars <- excel_min(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
