% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_mmult}
\alias{excel_mmult}
\title{Convert MMULT(A, B)
Performs AB = X and X are all matrices.}
\usage{
excel_mmult(A, B)
}
\arguments{
\item{A}{Matrix of size n x m}

\item{B}{MAtrix of size m x l}
}
\description{
Convert MMULT(A, B)
Performs AB = X and X are all matrices.
}
