% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-binom.R
\name{excel_negbinomdist}
\alias{excel_negbinomdist}
\title{Excel NEGBINOMDIST function}
\usage{
excel_negbinomdist(num_f, num_s, prob)
}
\arguments{
\item{num_f}{Number of failed trials}

\item{num_s}{Number of successful trials}

\item{prob}{Probability}
}
\value{
dnbinom
}
\description{
return correct norm distribution
}
\examples{
negbinomdist <- excel_negbinomdist(3,4,0.25)

}
