% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-normdist.R
\name{excel_normdist}
\alias{excel_normdist}
\title{Excel NORMDIST function}
\usage{
excel_normdist(x, mean, sd, cumulative)
}
\arguments{
\item{x}{list of random numbers}

\item{mean}{mean of distribution}

\item{sd}{standard deviation of distribution}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qnorm or dnorm
}
\description{
return correct norm distribution
}
\examples{
normdist <- excel_normdist(0.2,100,6,True)

}
