% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-normdist.R
\name{excel_normsdist}
\alias{excel_normsdist}
\title{Excel NORMSDIST function}
\usage{
excel_normsdist(x)
}
\arguments{
\item{x}{distribution value}
}
\value{
qnorm
}
\description{
Excel NORMSDIST function
}
\examples{
normsdist <- excel_normsdist(0.2)

}
