% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-poissondist.R
\name{excel_poisson}
\alias{excel_poisson}
\title{Excel POISSON function}
\usage{
excel_poisson(x, mean, cumulative)
}
\arguments{
\item{x}{Number of events}

\item{mean}{expectation value}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
ppois or dpois
}
\description{
return correct norm distribution
}
\examples{
poisson <- excel_poisson(10,10,True)

}
