% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_power}
\alias{excel_power}
\title{Convert POWER(a,b)}
\usage{
excel_power(a, b)
}
\arguments{
\item{a}{value}

\item{b}{exponent}
}
\value{
a^b
}
\description{
Convert POWER(a,b)
}
\examples{
powerSingleVars <- excel_power(2, 3)
}
