% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_product}
\alias{excel_product}
\title{Convert PRODUCT}
\usage{
excel_product(...)
}
\arguments{
\item{...}{list of inputs}
}
\value{
either single value product, or vector of products
}
\description{
Convert PRODUCT
}
\examples{
productSingleVars <- excel_product(0, 3, 5, 2.5, -10)
productListVars <- excel_product(list(2, 3, 5, 2.5, -10))
productMultiListVars <- excel_product(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))
}
