% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_quartile}
\alias{excel_quartile}
\title{Convert QUARTILE}
\usage{
excel_quartile(range, quart)
}
\arguments{
\item{range}{list of values}

\item{quart}{list of values}
}
\value{
returns the quartile determined by quart
}
\description{
Convert QUARTILE
}
\examples{
quartileListVars <- excel_quartile(list(2, 3, 5, 2.5, -10),1)
quartileMultiListVars <- excel_quartile(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)),1)

}
