% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-trig.R
\name{excel_radians}
\alias{excel_radians}
\title{Convert RADIANS}
\usage{
excel_radians(deg)
}
\arguments{
\item{deg}{angle in degrees}
}
\value{
angle in radians
}
\description{
Convert RADIANS
}
\examples{
radians <- excel_radians(180)
}
