% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-replacement.R
\name{excel_rand}
\alias{excel_rand}
\title{Excel RAND function}
\usage{
excel_rand(numberofruns, validate)
}
\arguments{
\item{numberofruns}{integer number of PSA runs}

\item{validate}{TRUE/FALSE}
}
\value{
vector of random numbers, or a vector of validation samples
}
\description{
Take number of random numbers, and the validation flag
}
\examples{
test <- excel_rand(numberofruns,validate)

}
