% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-string.R
\name{excel_right}
\alias{excel_right}
\title{Convert RIGHT(str, n)}
\usage{
excel_right(str, n)
}
\arguments{
\item{str}{text string}

\item{n}{number of characters to be captured}
}
\description{
Convert RIGHT(str, n)
}
\examples{
 rightstringOne<- excel_right("string",1)
 rightstringTwo<- excel_right("string",5)
 rightstringThree<- excel_right("string",3)

}
