% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_stdev}
\alias{excel_stdev}
\title{Convert STDEV(...)}
\usage{
excel_stdev(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
sample standard deviation
}
\description{
Convert STDEV(...)
}
\examples{
stdevSingleVars <- excel_stdev(0, 3, 5, 2.5, -10)
stdevListVars <- excel_stdev(list(2, 3, 5, 2.5, -10))
stdevMultiListVars <- excel_stdev(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
