% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_stdevp}
\alias{excel_stdevp}
\title{Convert STDEVP(...)}
\usage{
excel_stdevp(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
population standard deviation
}
\description{
Convert STDEVP(...)
}
\examples{
stdevpSingleVars <- excel_stdevp(0, 3, 5, 2.5, -10)
stdevpListVars <- excel_stdevp(list(2, 3, 5, 2.5, -10))
stdevpMultiListVars <- excel_stdevp(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
