% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_maths.R
\name{excel_sum}
\alias{excel_sum}
\title{Select which SUM should be used}
\usage{
excel_sum(...)
}
\arguments{
\item{...}{list of inputs}
}
\value{
either sum(c(...)) or Reduce('+',list(...))
}
\description{
Select the correct SUM type based on inputs.
If the inputs are a singular list, then call sum()
If inputs are a list of lists, call Reduce('+',list())
}
\examples{
sumSingleVars <- excel_sum(0, 3, 5, 2.5, -10)
sumListVars <- excel_sum(list(2, 3, 5, 2.5, -10))
sumMultiListVars <- excel_sum(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))
}
