% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-tdist.R
\name{excel_tdist}
\alias{excel_tdist}
\title{Excel TDIST function}
\usage{
excel_tdist(x, df, tails)
}
\arguments{
\item{x}{value}

\item{df}{degrees of freedom}

\item{tails}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
pt or dt
}
\description{
return correct norm distribution
}
\examples{
tdist <- excel_tdist(0.2,10,1)
tdist <- excel_tdist(0.2,10,2)

}
