% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_var}
\alias{excel_var}
\title{Convert VAR(...)}
\usage{
excel_var(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
sample variance of list
}
\description{
Convert VAR(...)
}
\examples{
varSingleVars <- excel_var(0, 3, 5, 2.5, -10)
varListVars <- excel_var(list(2, 3, 5, 2.5, -10))
varMultiListVars <- excel_var(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
