% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{excel_varp}
\alias{excel_varp}
\title{Convert VARP(...)}
\usage{
excel_varp(...)
}
\arguments{
\item{...}{list of values - allows non numeric characters}
}
\value{
population variance of list
}
\description{
Convert VARP(...)
}
\examples{
varpSingleVars <- excel_varp(0, 3, 5, 2.5, -10)
varpListVars <- excel_varp(list(2, 3, 5, 2.5, -10))
varpMultiListVars <- excel_varp(list(c(2, 3, 5), c(2.5 ,-10, 2), c(4, 3.3, 9)))

}
