% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-weibull.R
\name{excel_weibull}
\alias{excel_weibull}
\title{Excel WEIBULL function}
\usage{
excel_weibull(x, alpha, beta, cumulative)
}
\arguments{
\item{x}{value to evaluate}

\item{alpha}{shape}

\item{beta}{scale}

\item{cumulative}{TRUE/FALSE flag switching between cumulative distribution function, and probability mass function.}
}
\value{
qbinom or dbinom
}
\description{
Excel WEIBULL function
}
\examples{
weibull <- excel_weibull(0.2,100,6,True)

}
