% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_to_base}
\alias{formula_to_base}
\title{Convert a cell formula into expression of single cells rather than formulae}
\usage{
formula_to_base(cell_formula = "Engine!G5/Engine!G6", workbook_name)
}
\arguments{
\item{cell_formula}{string with the cell formula}

\item{workbook_name}{filename with location of workbook}
}
\value{
string with modified cell formula
}
\description{
Convert a cell formula into expression of single cells rather than formulae
}
\examples{
temp_output <- formula_to_base(cell_formula = "Engine!G5/Engine!G6", workbook_name = "test_workbook.xlsx")
}
