% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{generic_sd}
\alias{generic_sd}
\title{Generic sd}
\usage{
generic_sd(x, na.rm = FALSE, sample = TRUE)
}
\arguments{
\item{x}{vector of values to calculate standard deviation}

\item{na.rm}{TRUE/FALSE - ignore NA in list}

\item{sample}{TRUE/FALSE - when true returns sample standard deviation. When false returns population.}
}
\value{
Returns standard deviation of input vector
}
\description{
Generic sd
}
\examples{
popsd = generic_sd(c(1,2,3,4,5,6), sample = FALSE)
samplesd = generic_sd(c(1,2,3,4,5,6))
}
