% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-stats.R
\name{generic_var}
\alias{generic_var}
\title{Generic var}
\usage{
generic_var(x, na.rm = FALSE, sample = TRUE)
}
\arguments{
\item{x}{vector of values to calculate variance}

\item{na.rm}{TRUE/FALSE - ignore NA in list}

\item{sample}{TRUE/FALSE - when true returns sample variance. When false returns population.}
}
\value{
Returns variance of input vector
}
\description{
Generic var
}
\examples{
popvar = generic_var(c(1,2,3,4,5,6), sample = FALSE)
samplevar = generic_var(c(1,2,3,4,5,6))
}
