% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{is.blank}
\alias{is.blank}
\title{is.blank function}
\usage{
is.blank(x, false.triggers = FALSE)
}
\arguments{
\item{x}{the value to check}

\item{false.triggers}{values to trigger when false}
}
\value{
true/false
}
\description{
Check if value is a "blank" value
}
\examples{
sixth <- is.blank(6)
tenth <- is.blank(NA)

}
