% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank_1.R
\name{is_blank}
\alias{is_blank}
\title{Test if x is na or null
Tests if missing value and overcomes logical(0) errors when using is.na
Shared by Backlin 29-Oct-2013
https://stackoverflow.com/questions/19655579/a-function-that-returns-true-on-na-null-nan-in-r}
\usage{
is_blank(x, false_triggers = FALSE)
}
\arguments{
\item{x}{The scalar to test}

\item{false_triggers}{extra conditions}
}
\value{
Boolean indicating if x is blank
}
\description{
Test if x is na or null
Tests if missing value and overcomes logical(0) errors when using is.na
Shared by Backlin 29-Oct-2013
https://stackoverflow.com/questions/19655579/a-function-that-returns-true-on-na-null-nan-in-r
}
\examples{
temp1 <- is_blank(x = NA)
temp2 <- is_blank(x = 2)
}
