% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{PSA_output}
\alias{PSA_output}
\title{Write PSA to output}
\usage{
PSA_output(file, dec = ".", ...)
}
\arguments{
\item{file}{file path}

\item{dec}{allows user to determine their own radix symbol (default = ".")}

\item{...}{the N inputs}
}
\description{
Take N inputs and dumps them to screen
}
\examples{
PSA_output(c(1,0,1), c(1,2,3), c(3,4,5))

}
