% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{screen_dump}
\alias{screen_dump}
\title{Screen Dump}
\usage{
screen_dump(...)
}
\arguments{
\item{...}{the N inputs}
}
\value{
zero (0) for success, non-zero for failure
}
\description{
Take N inputs and dumps them to screen
}
\examples{
screen_dump(6, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)

}
